Revert "Discover tests with `ClassLoader` other than `ImageClassLoader` (#445)"
This reverts commit 298836496bc17f2f96514dde4eb455fc75ae24f6.
