+++ common/junit-platform-native/src/main/java/org/graalvm/junit/platform/NativeImageJUnitLauncher.java
@@ -43,8 +43,6 @@ package org.graalvm.junit.platform;
 
 import org.graalvm.nativeimage.ImageInfo;
 import org.graalvm.nativeimage.ImageSingletons;
-import org.graalvm.nativeimage.Platform;
-import org.graalvm.nativeimage.Platforms;
 import org.junit.platform.launcher.Launcher;
 import org.junit.platform.launcher.TestExecutionListener;
 import org.junit.platform.launcher.TestPlan;
@@ -61,17 +59,11 @@ public class NativeImageJUnitLauncher {
     static final String DEFAULT_OUTPUT_FOLDER = Paths.get("test-results-native").resolve("test").toString();
 
     final Launcher launcher;
-    TestPlan testPlan;
-    final TestsDiscoveryHelper testsDiscoveryHelper;
+    final TestPlan testPlan;
 
-    @Platforms(Platform.HOSTED_ONLY.class)
-    public NativeImageJUnitLauncher(TestsDiscoveryHelper testsDiscoveryHelper) {
-        this.testsDiscoveryHelper = testsDiscoveryHelper;
-        launcher = testsDiscoveryHelper.getLauncher();
-    }
-
-    private void discoverTests() {
-        testPlan = testsDiscoveryHelper.discoverTestPlan();
+    public NativeImageJUnitLauncher(Launcher launcher, TestPlan testPlan) {
+        this.launcher = launcher;
+        this.testPlan = testPlan;
     }
 
     public void registerTestExecutionListeners(TestExecutionListener testExecutionListener) {
@@ -123,8 +115,7 @@ public class NativeImageJUnitLauncher {
 
         PrintWriter out = new PrintWriter(System.out);
         NativeImageJUnitLauncher launcher = ImageSingletons.lookup(NativeImageJUnitLauncher.class);
-        //Discover the test plan at runtime.
-        launcher.discoverTests();
+
         if (!silent) {
             out.println("JUnit Platform on Native Image - report");
             out.println("----------------------------------------\n");
